restructure e2e test flows
