+++ tests/flows/governance/validator_set_sig.go
@@ -1,4 +1,4 @@
-package flows
+package governance
 
 import (
 	"context"
