+++ tests/flows/staking/erc20_token_staking.go
@@ -1,4 +1,4 @@
-package flows
+package staking
 
 import (
 	"context"
