+++ tests/flows/teleporter/add_fee_amount.go
@@ -1,4 +1,4 @@
-package flows
+package teleporter
 
 import (
 	"context"
