+++ tests/flows/teleporter/basic_send_receive.go
@@ -1,4 +1,4 @@
-package flows
+package teleporter
 
 import (
 	"context"
