+++ tests/flows/teleporter/deliver_to_nonexistent_contract.go
@@ -1,4 +1,4 @@
-package flows
+package teleporter
 
 import (
 	"context"
