+++ tests/flows/teleporter/deliver_to_wrong_chain.go
@@ -1,4 +1,4 @@
-package flows
+package teleporter
 
 import (
 	"context"
