+++ tests/flows/teleporter/insufficient_gas.go
@@ -1,4 +1,4 @@
-package flows
+package teleporter
 
 import (
 	"context"
