+++ tests/flows/teleporter/registry/check_upgrade_access.go
@@ -1,13 +1,15 @@
-package flows
+package registry
 
 import (
 	"context"
 	"math/big"
 
 	"github.com/ava-labs/subnet-evm/accounts/abi/bind"
+	"github.com/ava-labs/teleporter/tests/flows"
 	"github.com/ava-labs/teleporter/tests/interfaces"
 	"github.com/ava-labs/teleporter/tests/utils"
 	"github.com/ethereum/go-ethereum/crypto"
+
 	. "github.com/onsi/gomega"
 )
 
@@ -53,7 +55,7 @@ func CheckUpgradeAccess(network interfaces.Network) {
 	Expect(err).Should(BeNil())
 	_, err = testMessenger.PauseTeleporterAddress(nonOwnerOpts, teleporterAddress)
 	Expect(err).ShouldNot(BeNil())
-	Expect(err.Error()).Should(ContainSubstring(errTxReverted))
+	Expect(err.Error()).Should(ContainSubstring(flows.ErrTxReverted))
 
 	// Check that the teleporter address is not paused, because previous call should have failed
 	isPaused, err := testMessenger.IsTeleporterAddressPaused(&bind.CallOpts{}, teleporterAddress)
@@ -84,7 +86,7 @@ func CheckUpgradeAccess(network interfaces.Network) {
 	// Try to call unpauseTeleporterAddress from the previous owner account
 	_, err = testMessenger.UnpauseTeleporterAddress(ownerOpts, teleporterAddress)
 	Expect(err).ShouldNot(BeNil())
-	Expect(err.Error()).Should(ContainSubstring(errTxReverted))
+	Expect(err.Error()).Should(ContainSubstring(flows.ErrTxReverted))
 
 	// Make sure the teleporter address is still paused
 	isPaused, err = testMessenger.IsTeleporterAddressPaused(&bind.CallOpts{}, teleporterAddress)
