+++ tests/flows/teleporter/relay_message_twice.go
@@ -1,4 +1,4 @@
-package flows
+package teleporter
 
 import (
 	"context"
