+++ tests/flows/teleporter/relayer_modifies_message.go
@@ -1,4 +1,4 @@
-package flows
+package teleporter
 
 import (
 	"context"
