+++ tests/flows/teleporter/resubmit_altered_message.go
@@ -1,4 +1,4 @@
-package flows
+package teleporter
 
 import (
 	"context"
