+++ tests/flows/teleporter/retry_successful_execution.go
@@ -1,4 +1,4 @@
-package flows
+package teleporter
 
 import (
 	"context"
