+++ tests/flows/teleporter/send_specific_receipts.go
@@ -1,4 +1,4 @@
-package flows
+package teleporter
 
 import (
 	"bytes"
