+++ tests/flows/teleporter/teleporter_message_ids.go
@@ -1,7 +1,7 @@
 // (c) 2024, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package flows
+package teleporter
 
 import (
 	"math/big"
