+++ tests/flows/teleporter/unallowed_relayer.go
@@ -1,4 +1,4 @@
-package flows
+package teleporter
 
 import (
 	"context"
