+++ tests/flows/teleporter/validator_churn.go
@@ -1,4 +1,4 @@
-package flows
+package teleporter
 
 import (
 	"context"
